car/get_data_car/requirements.txt
requests
pymysql
pyjwt